updated file structure for transfer
